tvb_framework/requirements.txt
numba
numpy
scipy
networkx
scikit-learn
scikit-image
cython
numexpr
psutil
pytest
pytest-cov
pytest-xdist
pytest-benchmark
pytest-mock
matplotlib==3.5.3
h5py
typing
BeautifulSoup4
subprocess32
python-keycloak
mako
pyAesCrypt
pyunicore
formencode
flask
flask-restx
jinja2
nibabel>=4.0
sqlalchemy<2.0
alembic
allensdk
werkzeug
gevent
cherrypy
MarkupSafe
decorator
tables==3.7.0
pandas
requests
simplejson
six
siibra==0.3a27
bctpy
kubernetes
watchdog